Adding a preprocess module.
