+++ tests/test_compute.py
@@ -14,7 +14,6 @@ import pytest
 
 import numpy as np
 
-from matrixprofile.compute import preprocess
 from matrixprofile import compute
 
 import matrixprofile
@@ -131,13 +130,3 @@ def test_compute_mp_invalid_windows():
         compute(ts, windows=w)
         assert 'Compute requires all window sizes to be greater than 3!' \
             in str(excinfo.value)
-
-
-def test_preprocess_data():
-    ts = np.array([np.nan, np.inf, np.inf, np.nan, np.inf, 2, 3, 2, 3, 1, 2, 3, 4, 2,
-                   np.nan, np.inf, 4, 2, 3, 4, 5, 6, 7, 8, 3, 4, 2, 3, 4, 5, 6, 7, 6,
-                   5, 4, 3, np.nan, np.nan, np.inf, np.nan, np.inf, np.nan])
-
-    ts = preprocess(ts, 4)
-    assert(np.any(np.isnan(ts)) == False)
-    assert(np.any(np.isinf(ts)) == False)
